requirements.txt
langchain-groq==0.1.5
langgraph==0.0.66
sentence-transformers==2.2.2
chromadb==0.4.22
streamlit==1.28.1
python-dotenv==1.0.0
ragas==0.0.22
datasets
torch
numpy
pandas
